+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/connection/DdlBatchTest.java
@@ -156,11 +156,9 @@ public class DdlBatchTest {
     when(singleUse.executeQuery(statement)).thenReturn(resultSet);
     when(dbClient.singleUse()).thenReturn(singleUse);
     DdlBatch batch = createSubject(createDefaultMockDdlClient(), dbClient);
-    assertThat(
-        batch
-            .executeQuery(parsedStatement, AnalyzeMode.NONE, InternalMetadataQuery.INSTANCE)
-            .hashCode(),
-        is(equalTo(resultSet.hashCode())));
+    ResultSet result =
+        batch.executeQuery(parsedStatement, AnalyzeMode.NONE, InternalMetadataQuery.INSTANCE);
+    assertThat(result.hashCode(), is(equalTo(resultSet.hashCode())));
   }
 
   @Test